feat: Detach t_sent add functions from sent_create.c file to a new file sent_add.c file
