+++ src/t_sent/sent_create.c
@@ -6,7 +6,7 @@
 /*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/28 16:07:12 by sanghupa          #+#    #+#             */
-/*   Updated: 2023/08/07 10:27:25 by sanghupa         ###   ########.fr       */
+/*   Updated: 2023/08/24 16:50:10 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,43 +88,4 @@ t_sent	*sent_new(char *p_unit, int is_redir, int is_pipe)
 	this->prev = NULL;
 	this->next = NULL;
 	return (this);
-}
-
-void	sent_addfront(t_sent *sent[], t_sent *new)
-{
-	t_sent	*head;
-
-	head = *sent;
-	if (!sent || !new)
-		return ;
-	if (!head)
-	{
-		*sent = new;
-		new->prev = NULL;
-		new->next = NULL;
-		return ;
-	}
-	head->prev = new;
-	new->next = head;
-	*sent = new;
-	new->prev = NULL;
-}
-
-void	sent_addback(t_sent *sent[], t_sent *new)
-{
-	t_sent	*tail;
-
-	if (!sent || !new)
-		return ;
-	tail = sent_getlast(*sent);
-	if (!tail)
-	{
-		*sent = new;
-		new->prev = NULL;
-		new->next = NULL;
-		return ;
-	}
-	tail->next = new;
-	new->prev = tail;
-	new->next = NULL;
 }
